+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/pipeline/version/handler/PipelineDraftReleaseHandler.kt
@@ -44,13 +44,12 @@ import com.tencent.devops.process.pojo.pipeline.PipelineYamlFileReleaseReqSource
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionCreateContext
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionGenerator
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionPersistenceService
-import com.tencent.devops.process.yaml.PipelineYamlCommonService
-import com.tencent.devops.process.yaml.PipelineYamlFacadeService
+import com.tencent.devops.process.yaml.PipelineYamlReleaseService
+import com.tencent.devops.process.yaml.PipelineYamlService
 import jakarta.ws.rs.core.Response
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.context.annotation.Lazy
 import org.springframework.stereotype.Service
 
 /**
@@ -64,9 +63,8 @@ class PipelineDraftReleaseHandler @Autowired constructor(
     private val dslContext: DSLContext,
     private val pipelineResourceVersionDao: PipelineResourceVersionDao,
     private val pipelineResourceDao: PipelineResourceDao,
-    @Lazy
-    private val pipelineYamlFacadeService: PipelineYamlFacadeService,
-    private val pipelineYamlCommonService: PipelineYamlCommonService
+    private val pipelineYamlService: PipelineYamlService,
+    private val pipelineYamlReleaseService: PipelineYamlReleaseService
 ) : PipelineVersionCreateHandler {
     override fun support(context: PipelineVersionCreateContext): Boolean {
         return context.versionAction == PipelineVersionAction.RELEASE_DRAFT
@@ -151,19 +149,10 @@ class PipelineDraftReleaseHandler @Autowired constructor(
                 updateBuildNo = true
             }
         }
-        // 检查推送参数
-        enablePac.takeIf { it }?.let {
-            pipelineYamlCommonService.checkPushParam(
-                projectId = projectId,
-                pipelineId = pipelineId,
-                content = pipelineResourceWithoutVersion.yaml!!,
-                repoHashId = yamlFileInfo!!.repoHashId,
-                filePath = yamlFileInfo.filePath,
-                targetAction = targetAction!!,
-                versionName = resourceOnlyVersion.versionName,
-                targetBranch = targetBranch
-            )
-        }
+        pipelineYamlReleaseService.validateReleaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion
+        )
         if (pipelineResourceWithoutVersion.status == VersionStatus.RELEASED) {
             pipelineVersionPersistenceService.releaseDraft2ReleaseVersion(
                 context = this,
@@ -177,15 +166,13 @@ class PipelineDraftReleaseHandler @Autowired constructor(
         }
 
         // 推送文件
-        val yamlFileReleaseResult = enablePac.takeIf { it }?.let {
-            pipelineVersionPersistenceService.releaseYamlFile(
-                context = this,
-                resourceOnlyVersion = resourceOnlyVersion,
-                source = PipelineYamlFileReleaseReqSource.PIPELINE
-            )
-        }
+        val yamlFileReleaseResult = pipelineYamlReleaseService.releaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion,
+            source = PipelineYamlFileReleaseReqSource.PIPELINE
+        )
 
-        val yamlInfo = pipelineYamlFacadeService.getPipelineYamlInfo(
+        val yamlInfo = pipelineYamlService.getPipelineYamlVo(
             projectId = projectId,
             pipelineId = pipelineId,
             version = resourceOnlyVersion.version
