+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/pipeline/version/handler/PipelineTemplateInstanceHandler.kt
@@ -41,7 +41,7 @@ import com.tencent.devops.process.service.PipelineRemoteAuthService
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionCreateContext
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionGenerator
 import com.tencent.devops.process.service.pipeline.version.PipelineVersionPersistenceService
-import com.tencent.devops.process.yaml.PipelineYamlCommonService
+import com.tencent.devops.process.yaml.PipelineYamlReleaseService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
@@ -51,7 +51,7 @@ class PipelineTemplateInstanceHandler @Autowired constructor(
     private val redisOperation: RedisOperation,
     private val pipelineVersionGenerator: PipelineVersionGenerator,
     private val pipelineVersionPersistenceService: PipelineVersionPersistenceService,
-    private val pipelineYamlCommonService: PipelineYamlCommonService,
+    private val pipelineYamlReleaseService: PipelineYamlReleaseService,
     private val pipelineRemoteAuthService: PipelineRemoteAuthService
 ) : PipelineVersionCreateHandler {
     override fun support(context: PipelineVersionCreateContext) =
@@ -120,20 +120,10 @@ class PipelineTemplateInstanceHandler @Autowired constructor(
             )
         }
 
-        // 检查推送参数
-        enablePac.takeIf { it }?.let {
-            pipelineYamlCommonService.checkPushParam(
-                projectId = projectId,
-                pipelineId = pipelineId,
-                content = pipelineResourceWithoutVersion.yaml!!,
-                repoHashId = yamlFileInfo!!.repoHashId,
-                filePath = yamlFileInfo.filePath,
-                targetAction = targetAction!!,
-                versionName = resourceOnlyVersion.versionName,
-                targetBranch = targetBranch
-            )
-        }
-
+        pipelineYamlReleaseService.validateReleaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion
+        )
         when {
             pipelineInfo == null -> {
                 pipelineVersionPersistenceService.initializePipeline(
@@ -166,13 +156,11 @@ class PipelineTemplateInstanceHandler @Autowired constructor(
         }
 
         // 推送文件
-        val yamlFileReleaseResult = enablePac.takeIf { it }?.let {
-            pipelineVersionPersistenceService.releaseYamlFile(
-                context = this,
-                resourceOnlyVersion = resourceOnlyVersion,
-                source = PipelineYamlFileReleaseReqSource.TEMPLATE_INSTANCE
-            )
-        }
+        val yamlFileReleaseResult = pipelineYamlReleaseService.releaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion,
+            source = PipelineYamlFileReleaseReqSource.TEMPLATE_INSTANCE
+        )
 
         return DeployPipelineResult(
             pipelineId = pipelineId,
@@ -206,7 +194,6 @@ class PipelineTemplateInstanceHandler @Autowired constructor(
             pipelineRemoteAuthService.generateAuth(pipelineId, projectId, userId)
         }
     }
-
     companion object {
         private val logger = LoggerFactory.getLogger(PipelineTemplateInstanceHandler::class.java)
     }
