+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/version/hander/PipelineTemplateDraftReleaseHandler.kt
@@ -34,20 +34,15 @@ import com.tencent.devops.common.pipeline.enums.VersionStatus
 import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.process.constant.ProcessMessageCode
 import com.tencent.devops.process.pojo.pipeline.DeployTemplateResult
-import com.tencent.devops.process.pojo.pipeline.PipelineYamlFileReleaseReq
-import com.tencent.devops.process.pojo.pipeline.PipelineYamlFileReleaseReqSource
-import com.tencent.devops.process.pojo.pipeline.PipelineYamlFileReleaseResult
-import com.tencent.devops.process.pojo.template.v2.PTemplateResourceOnlyVersion
 import com.tencent.devops.process.service.template.v2.PipelineTemplateGenerator
 import com.tencent.devops.process.service.template.v2.PipelineTemplateModelLock
 import com.tencent.devops.process.service.template.v2.PipelineTemplatePersistenceService
 import com.tencent.devops.process.service.template.v2.PipelineTemplateResourceService
 import com.tencent.devops.process.service.template.v2.PipelineTemplateSettingService
 import com.tencent.devops.process.service.template.v2.version.PipelineTemplateVersionCreateContext
-import com.tencent.devops.process.yaml.PipelineYamlFacadeService
+import com.tencent.devops.process.yaml.PipelineYamlReleaseService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.context.annotation.Lazy
 import org.springframework.stereotype.Service
 
 /**
@@ -60,7 +55,7 @@ class PipelineTemplateDraftReleaseHandler @Autowired constructor(
     private val pipelineTemplateResourceService: PipelineTemplateResourceService,
     private val pipelineTemplateSettingService: PipelineTemplateSettingService,
     private val redisOperation: RedisOperation,
-    @Lazy private val pipelineYamlFacadeService: PipelineYamlFacadeService
+    private val pipelineYamlReleaseService: PipelineYamlReleaseService
 ) : PipelineTemplateVersionCreateHandler {
     override fun support(context: PipelineTemplateVersionCreateContext) =
         context.versionAction == PipelineVersionAction.RELEASE_DRAFT
@@ -134,7 +129,10 @@ class PipelineTemplateDraftReleaseHandler @Autowired constructor(
             targetAction = targetAction,
             targetBranch = branchName
         )
-        validateReleaseYamlFile(resourceOnlyVersion = resourceOnlyVersion)
+        pipelineYamlReleaseService.validateReleaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion
+        )
         if (versionStatus == VersionStatus.RELEASED) {
             pipelineTemplatePersistenceService.releaseDraft2ReleaseVersion(
                 context = this,
@@ -148,7 +146,10 @@ class PipelineTemplateDraftReleaseHandler @Autowired constructor(
         }
 
         // 发布yaml文件
-        val yamlFileReleaseResult = releaseYamlFile(resourceOnlyVersion = resourceOnlyVersion)
+        val yamlFileReleaseResult = pipelineYamlReleaseService.releaseYamlFile(
+            context = this,
+            resourceOnlyVersion = resourceOnlyVersion
+        )
         return DeployTemplateResult(
             projectId = projectId,
             userId = userId,
@@ -160,60 +161,6 @@ class PipelineTemplateDraftReleaseHandler @Autowired constructor(
             versionName = resourceOnlyVersion.versionName,
             targetUrl = yamlFileReleaseResult?.pullRequestUrl,
             versionAction = versionAction
-        )
-    }
-
-    private fun PipelineTemplateVersionCreateContext.releaseYamlFile(
-        resourceOnlyVersion: PTemplateResourceOnlyVersion
-    ): PipelineYamlFileReleaseResult? {
-        if (!enablePac) {
-            return null
-        }
-        val yamlFileReleaseReq = PipelineYamlFileReleaseReq(
-            userId = userId,
-            projectId = projectId,
-            pipelineId = templateId,
-            pipelineName = pipelineTemplateInfo.name,
-            version = resourceOnlyVersion.version.toInt(),
-            versionName = resourceOnlyVersion.versionName,
-            repoHashId = yamlFileInfo!!.repoHashId,
-            filePath = yamlFileInfo.filePath,
-            content = pTemplateResourceWithoutVersion.yaml!!,
-            commitMessage = pTemplateResourceWithoutVersion.description
-                ?: "update template ${pipelineTemplateInfo.name}",
-            targetAction = targetAction!!,
-            targetBranch = branchName,
-            source = PipelineYamlFileReleaseReqSource.TEMPLATE
-        )
-        return pipelineYamlFacadeService.releaseYamlFile(
-            yamlFileReleaseReq = yamlFileReleaseReq
-        )
-    }
-
-    private fun PipelineTemplateVersionCreateContext.validateReleaseYamlFile(
-        resourceOnlyVersion: PTemplateResourceOnlyVersion
-    ) {
-        if (!enablePac) {
-            return
-        }
-        val yamlFileReleaseReq = PipelineYamlFileReleaseReq(
-            userId = userId,
-            projectId = projectId,
-            pipelineId = templateId,
-            pipelineName = pipelineTemplateInfo.name,
-            version = resourceOnlyVersion.version.toInt(),
-            versionName = resourceOnlyVersion.versionName,
-            repoHashId = yamlFileInfo!!.repoHashId,
-            filePath = yamlFileInfo.filePath,
-            content = pTemplateResourceWithoutVersion.yaml!!,
-            commitMessage = pTemplateResourceWithoutVersion.description
-                ?: "update template ${pipelineTemplateInfo.name}",
-            targetAction = targetAction!!,
-            targetBranch = branchName,
-            source = PipelineYamlFileReleaseReqSource.TEMPLATE
-        )
-        pipelineYamlFacadeService.validateReleaseYamlFile(
-            yamlFileReleaseReq = yamlFileReleaseReq
         )
     }
